use explicit 32bit data type in ANSC_IPV4_ADDRESS etc
An IPv4 address is always 32 bits. Using "unsigned long" to hold it
leads to compiler warnings etc (and then various unnecessary casts
etc in the code to avoid them). Using the correct sized data type is
a better solution.

+++ cm_hal.h
@@ -68,6 +68,7 @@
 #ifndef __CM_HAL_H__
 #define __CM_HAL_H__
 
+#include <stdint.h>
 #include <sys/time.h>
 
 /**********************************************************************
@@ -117,7 +118,7 @@ extern "C"{
          union                                                                              \
          {                                                                                  \
             unsigned char           Dot[IPV4_ADDRESS_SIZE];                                 \
-            unsigned long           Value;                                                  \
+            uint32_t                Value;                                                  \
          }
 #endif
